requirements.txt
flask
flask-cors
librosa
numpy
onnx
onnxruntime
scipy
speechmatics-python
spacy
translate
werkzeug
gunicorn